requirements.txt
google-adk>=1.28.1
litellm>=1.52.0
anthropic>=0.39.0
boto3>=1.35.0
botocore>=1.35.0
mcp>=1.27.0
httpx>=0.27.0
pydantic>=2.0.0
uvicorn>=0.30.0
starlette>=0.40.0
fastapi>=0.115.0
pytest>=8.0.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
langsmith>=0.1.95
